+++ PvPStatsClassic/EventHandlers.lua
@@ -446,7 +446,7 @@ local function HandleUnitDiedEvent(destGUID, destName)
     end
 end
 
-local function HandleComatLogEventPetDamage(combatEvent, sourceGUID, sourceName, destGUID, destName, param1, param4)
+local function HandleCombatLogEventPetDamage(combatEvent, sourceGUID, sourceName, destGUID, destName, param1, param4)
     if IsPetGUID(sourceGUID) and destGUID then
         local damageAmount = 0
 
@@ -519,7 +519,7 @@ local function HandleCombatLogEvent()
     local isValidTarget = PSC_IsValidTarget(destFlags, destGUID)
 
     if isValidTarget then
-        HandleComatLogEventPetDamage(combatEvent, sourceGUID, sourceName, destGUID, destName, param1, param4)
+        HandleCombatLogEventPetDamage(combatEvent, sourceGUID, sourceName, destGUID, destName, param1, param4)
         HandleCombatLogPlayerDamage(combatEvent, sourceGUID, sourceName, destGUID, destName, destFlags, param1, param4)
     end
 
@@ -656,6 +656,11 @@ local function HandlePlayerEnteringWorld()
         PSC_DB.WhatsNewPopupVersion = "v1.0"
     end
 
+    if PSC_DB.PlayerInfoCache == nil then
+        PSC_DB.PlayerInfoCache = {}
+    end
+    PSC_GetAndStorePlayerInfoFromUnit("player", true)
+
     PSC_MigratePlayerInfoCache()
     PSC_MigratePlayerInfoToEnglish()
     PSC_MigrateKillKeys()
@@ -664,6 +669,7 @@ local function HandlePlayerEnteringWorld()
     PSC_InitializePlayerKillCounts()
     PSC_InitializeLeaderboardCache()
     PSC_InitializePlayerLossCounts()
+
     PSC_UpdateMinimapButtonPosition()
     PSC_SetupMouseoverTooltip()
     PSC_InCombat = UnitAffectingCombat("player")
@@ -687,7 +693,7 @@ local function HandlePlayerEnteringWorld()
     local currentVersion = PSC_GetAddonVersion()
     if PSC_DB.WhatsNewPopupVersion ~= currentVersion then
         local title = "PvP Stats v" .. currentVersion .. " - What's new:"
-        local message = "-Added Arena achievements.\n\n-Fixed \"Unknown\" zone name for kills/deaths in Arena.\n\n-Fixed double-kills for priests with Spirit of Redemption talent.\n\n-You can now import your data from older clients like Classic Era. Use the link below.\n\nEnjoy!"
+        local message = "-Fixed \"You aren't in a party\" messages in BGs\n\n-Leaderboard now has option to show your own alts\n\n-Added Bonus achievement to kill the Defias Traitor 250 times\n\n-You can import your data from older clients like Classic Era. Use the link below.\n\nEnjoy!"
         local dataImportGuideUrl = "https://github.com/randomdude163/WoWClassic_PvPStats/wiki/How-to-import-data-from-other-WoW-clients-(like-Classic-Era)"
         PSC_ShowWhatsNewPopup(title, message, function()
             PSC_DB.WhatsNewPopupShown = true
@@ -764,6 +770,7 @@ function PSC_RegisterEvents()
         elseif event == "PLAYER_REGEN_ENABLED" then
             HandlePlayerRegenEnabled()
         elseif event == "PLAYER_LOGOUT" then
+            PSC_GetAndStorePlayerInfoFromUnit("player", true)
             PSC_CleanupPlayerInfoCache()
         elseif event == "ZONE_CHANGED_NEW_AREA" then
             PSC_CheckBattlegroundStatus()